requirements.txt
black==19.10b0
django-db-geventpool==3.2.1
django-environ==0.4.5
djangorestframework==3.11.0
django==3.0.5
gevent==1.4.0
gunicorn==20.0.4
ipython==7.13.0
psycopg2-binary==2.8.4
pylint-django==2.0.14
pylint==2.4.4
pytz==2019.3
requests==2.23.0
rope==0.16.0
sentry-sdk==0.14.3
whitenoise==5.0.1
